Decouple sidecar from Solana
