--- sidecar/solanarpc/BUILD.bazel
@@ -1,39 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
-load("//.bazel:defs.bzl", "go_test")
-
-go_library(
-    name = "solanarpc_lib",
-    srcs = [
-        "main.go",
-        "rpc.go",
-        "server.go",
-        "verifier.go",
-    ],
-    importpath = "github.com/ava-labs/avalanchego/sidecar/solanarpc",
-    visibility = ["//visibility:private"],
-    deps = [
-        "//network/p2p/oracle",
-        "@com_github_mr_tron_base58//:base58",
-    ],
-)
-
-go_binary(
-    name = "solanarpc",
-    embed = [":solanarpc_lib"],
-    visibility = ["//visibility:public"],
-)
-
-go_test(
-    name = "solanarpc_test",
-    srcs = [
-        "integration_test.go",
-        "server_test.go",
-        "verifier_test.go",
-    ],
-    embed = [":solanarpc_lib"],
-    deps = [
-        "//network/p2p/oracle",
-        "@com_github_mr_tron_base58//:base58",
-        "@com_github_stretchr_testify//require",
-    ],
-)
