+++ sidecar/solanarpc/integration_test.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2019, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package main
+package solanarpc
 
 import (
 	"bytes"
