+++ sidecar/solanarpc/verifier.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2019, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package main
+package solanarpc
 
 import (
 	"bytes"
@@ -14,11 +14,6 @@ import (
 
 	"github.com/ava-labs/avalanchego/network/p2p/oracle"
 )
-
-// errSourceChainUnavailable is returned when the Solana RPC cannot be reached.
-// server.go uses errors.Is against this sentinel to map the failure to
-// codes.Unavailable rather than codes.InvalidArgument.
-var errSourceChainUnavailable = errors.New("source chain unavailable")
 
 // SolanaVerifier verifies OracleMessages by querying the Solana RPC.
 type SolanaVerifier struct {
@@ -70,7 +65,7 @@ func (v *SolanaVerifier) Verify(ctx context.Context, msg *oracle.OracleMessage, 
 
 	tx, err := v.client.getTransaction(ctx, sig)
 	if err != nil {
-		return fmt.Errorf("%w: getTransaction RPC call failed: %w", errSourceChainUnavailable, err)
+		return fmt.Errorf("%w: getTransaction RPC call failed: %w", oracle.ErrSourceUnavailable, err)
 	}
 	if tx == nil {
 		return fmt.Errorf("transaction not found for signature %s", sig)
